refactor : notice 페이지 관련 타입 분리 및 관련 영역 적용
