+++ src/components/NoticeKeyword/index.tsx
@@ -3,12 +3,12 @@ import 'moment/locale/ko';
 import moment from 'moment';
 import { useRecoilValue } from 'recoil';
 
-import { NoticeKeywordDataType } from '@api/notice';
 import NoticeDeleteBtn from '@components/NoticeDeleteBtn';
 import * as S from '@components/NoticeKeyword/NoticeKeyword.style';
 import ImgContainer from '@components/common/ImgContainer';
 import { newShareEnrolledMention } from '@constants/mentions';
 import { deleteModeState } from '@store/notice';
+import { NoticeKeywordDataType } from '@type/notice';
 
 type NoticeKeywordPropsType = {
   contents: NoticeKeywordDataType[];
